requirements.txt
streamlit
altair<5.0
langchain
langchain_groq
langchain_openai
langchain_core
langchain_community
faiss-cpu
python-dotenv